Query-FROM-CSV/requirements.txt
streamlit
pandas
langchain
langchain_groq
matplotlib
python-dotenv